requirements.txt
playwright
pytest-playwright
pytest-dependency
